renaming stoppedTracks->isolatedTracks & merging collections
