+++ DataFormats/PatCandidates/src/classes_def_objects.xml
@@ -105,11 +105,7 @@
    <version ClassVersion="10" checksum="2244564938"/>
   </class>
 
-  <class name="pat::IsolatedTrack" ClassVersion="3">
-   <version ClassVersion="3" checksum="550880582"/>
-  </class>
-
-  <class name="pat::StoppedTrack" />
+  <class name="pat::IsolatedTrack" />
 
   <class name="pat::PFIsolation" ClassVersion="3">
     <version ClassVersion="3" checksum="2961562419" />
@@ -425,7 +421,6 @@
   <class name="std::vector<pat::Conversion>" />
   <class name="std::vector<pat::PackedCandidate>"/>
   <class name="std::vector<pat::IsolatedTrack>"/>
-  <class name="std::vector<pat::StoppedTrack>"/>
   <class name="std::vector<pat::PackedGenParticle>"/>
 
   <!-- PAT Object Collection Iterators -->
@@ -461,7 +456,6 @@
   <class name="edm::Wrapper<std::vector<pat::Conversion> >" />
   <class name="edm::Wrapper<std::vector<pat::PackedCandidate> >"/>
   <class name="edm::Wrapper<std::vector<pat::IsolatedTrack> >"/>
-  <class name="edm::Wrapper<std::vector<pat::StoppedTrack> >"/>
   <class name="edm::Wrapper<std::vector<pat::PackedGenParticle> >"/>
 
   <!-- PAT Object References -->
