+++ PhysicsTools/PatAlgos/plugins/PATIsolatedTrackProducer.cc
@@ -1,12 +1,6 @@
 #include <string>
-#include "DataFormats/PatCandidates/interface/PackedCandidate.h"
-#include "DataFormats/PatCandidates/interface/IsolatedTrack.h"
-#include "DataFormats/PatCandidates/interface/PFIsolation.h"
-#include "DataFormats/Candidate/interface/CandidateFwd.h"
-#include "FWCore/Framework/interface/global/EDProducer.h"
-#include "DataFormats/Common/interface/View.h"
-#include "DataFormats/Math/interface/deltaR.h"
-#include "DataFormats/Common/interface/RefToPtr.h"
+
+#include "FWCore/Framework/interface/EDProducer.h"
 #include "FWCore/Framework/interface/ESHandle.h"
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/EventSetup.h"
@@ -14,79 +8,356 @@
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/Utilities/interface/Exception.h"
 
+#include "DataFormats/ParticleFlowCandidate/interface/PFCandidate.h"
+
+#include "DataFormats/PatCandidates/interface/PackedCandidate.h"
+#include "DataFormats/PatCandidates/interface/IsolatedTrack.h"
+#include "DataFormats/PatCandidates/interface/PFIsolation.h"
+#include "DataFormats/Candidate/interface/CandidateFwd.h"
+#include "DataFormats/Common/interface/View.h"
+#include "DataFormats/Common/interface/RefToPtr.h"
+#include "DataFormats/Math/interface/deltaR.h"
+#include "DataFormats/DetId/interface/DetId.h"
+#include "DataFormats/HcalDetId/interface/HcalDetId.h"
+#include "DataFormats/JetReco/interface/CaloJet.h"
+#include "DataFormats/TrackReco/interface/Track.h"
+#include "DataFormats/TrackReco/interface/TrackExtraFwd.h"
+#include "DataFormats/TrackReco/interface/DeDxData.h"
+#include "DataFormats/TrackReco/interface/DeDxHitInfo.h"
+#include "RecoTracker/DeDx/interface/DeDxTools.h"
+
+#include "TrackingTools/TrackAssociator/interface/TrackDetectorAssociator.h"
+#include "TrackingTools/TrackAssociator/interface/TrackAssociatorParameters.h"
+#include "TrackingTools/TrajectoryState/interface/TrajectoryStateTransform.h"
+
+#include "CondFormats/HcalObjects/interface/HcalChannelQuality.h"
+#include "CondFormats/HcalObjects/interface/HcalChannelStatus.h"
+#include "CondFormats/EcalObjects/interface/EcalChannelStatus.h"
+#include "CondFormats/DataRecord/interface/HcalChannelQualityRcd.h"
+#include "CondFormats/DataRecord/interface/EcalChannelStatusRcd.h"
+
+#include "MagneticField/Engine/interface/MagneticField.h"
+
 namespace pat {
 
-    class PATIsolatedTrackProducer : public edm::global::EDProducer<> {
-        public:
-            explicit PATIsolatedTrackProducer(const edm::ParameterSet&);
-            ~PATIsolatedTrackProducer();
+    class PATIsolatedTrackProducer : public edm::EDProducer {
+      public:
+          typedef pat::IsolatedTrack::LorentzVector LorentzVector;
 
-        virtual void produce(edm::StreamID, edm::Event&, const edm::EventSetup&) const override;
+          explicit PATIsolatedTrackProducer(const edm::ParameterSet&);
+          ~PATIsolatedTrackProducer();
 
-        private: 
-            
-            const edm::EDGetTokenT<pat::PackedCandidateCollection>    pc_;
-            const float pT_cut;  // only save cands with pT>pT_cut
-            const float dR_cut;  // isolation radius
-            const float dZ_cut;  // save if either from PV or |dz|<dZ_cut
-            const std::vector<double> miniIsoParams;
-            const float absIso_cut;  // save if ANY of absIso, relIso, or miniRelIso pass the cuts 
-            const float relIso_cut;
-            const float miniRelIso_cut;
+          virtual void produce(edm::Event&, const edm::EventSetup&);
+        
+          // compute iso/miniiso
+          void getIsolation(const LorentzVector& p4, const pat::PackedCandidateCollection* pc, int pc_idx,
+                            pat::PFIsolation &iso, pat::PFIsolation &miniiso);
+
+          float getDeDx(const reco::DeDxHitInfo *hitInfo, bool doPixel, bool doStrip);
+
+          TrackDetMatchInfo getTrackDetMatchInfo(const edm::Event&, const edm::EventSetup&, const reco::Track&);
+
+          void getCaloJetEnergy(const LorentzVector&, const reco::CaloJetCollection*, float&, float&);
+
+      private:             
+          const edm::EDGetTokenT<pat::PackedCandidateCollection>    pc_;
+          const edm::EDGetTokenT<pat::PackedCandidateCollection>    lt_;
+          const edm::EDGetTokenT<reco::TrackCollection>    gt_;
+          const edm::EDGetTokenT<edm::Association<pat::PackedCandidateCollection> > gt2pc_;
+          const edm::EDGetTokenT<edm::Association<pat::PackedCandidateCollection> > gt2lt_;
+          const edm::EDGetTokenT<edm::Association<reco::PFCandidateCollection> > pc2pf_;
+          const edm::EDGetTokenT<reco::CaloJetCollection> caloJets_;
+          const edm::EDGetTokenT<edm::ValueMap<reco::DeDxData> > gt2dedx_;
+          const edm::EDGetTokenT<reco::DeDxHitInfoAss> gt2dedxHitInfo_;
+          const float pT_cut_;  // only save cands with pT>pT_cut_
+          const float pT_cut_noIso_;  // above this pT, don't apply any iso cut
+          const float pfIsolation_DR_;  // isolation radius
+          const float pfIsolation_DZ_;  // used in determining if pfcand is from PV or PU
+          const std::vector<double> miniIsoParams_;
+          const float absIso_cut_;  // save if ANY of absIso, relIso, or miniRelIso pass the cuts 
+          const float relIso_cut_;
+          const float miniRelIso_cut_;
+          const float caloJet_DR_;  // save energy of nearest calojet within caloJet_DR_
+
+          TrackDetectorAssociator trackAssociator_;
+          TrackAssociatorParameters trackAssocParameters_;
     };
 }
 
 pat::PATIsolatedTrackProducer::PATIsolatedTrackProducer(const edm::ParameterSet& iConfig) :
-  pc_(consumes<pat::PackedCandidateCollection>(iConfig.getParameter<edm::InputTag>("packedCandidates"))),
-  pT_cut(iConfig.getParameter<double>("pT_cut")),
-  dR_cut(iConfig.getParameter<double>("dR_cut")),
-  dZ_cut(iConfig.getParameter<double>("dZ_cut")),
-  miniIsoParams(iConfig.getParameter<std::vector<double> >("miniIsoParams")),
-  absIso_cut(iConfig.getParameter<double>("absIso_cut")),
-  relIso_cut(iConfig.getParameter<double>("relIso_cut")),
-  miniRelIso_cut(iConfig.getParameter<double>("miniRelIso_cut"))
+  pc_(consumes<pat::PackedCandidateCollection>(iConfig.getParameter<edm::InputTag>("packedPFCandidates"))),
+  lt_(consumes<pat::PackedCandidateCollection>(iConfig.getParameter<edm::InputTag>("lostTracks"))),
+  gt_(consumes<reco::TrackCollection>(iConfig.getParameter<edm::InputTag>("generalTracks"))),
+  gt2pc_(consumes<edm::Association<pat::PackedCandidateCollection> >(iConfig.getParameter<edm::InputTag>("packedPFCandidates"))),
+  gt2lt_(consumes<edm::Association<pat::PackedCandidateCollection> >(iConfig.getParameter<edm::InputTag>("lostTracks"))),
+  pc2pf_(consumes<edm::Association<reco::PFCandidateCollection> >(iConfig.getParameter<edm::InputTag>("packedPFCandidates"))),
+  caloJets_(consumes<reco::CaloJetCollection>(iConfig.getParameter<edm::InputTag>("caloJets"))),
+  gt2dedx_(consumes<edm::ValueMap<reco::DeDxData> >(iConfig.getParameter<edm::InputTag>("dEdxInfo"))),
+  gt2dedxHitInfo_(consumes<reco::DeDxHitInfoAss>(iConfig.getParameter<edm::InputTag>("dEdxHitInfo"))),
+  pT_cut_(iConfig.getParameter<double>("pT_cut")),
+  pT_cut_noIso_(iConfig.getParameter<double>("pT_cut_noIso")),
+  pfIsolation_DR_(iConfig.getParameter<double>("pfIsolation_DR")),
+  pfIsolation_DZ_(iConfig.getParameter<double>("pfIsolation_DZ")),
+  miniIsoParams_(iConfig.getParameter<std::vector<double> >("miniIsoParams")),
+  absIso_cut_(iConfig.getParameter<double>("absIso_cut")),
+  relIso_cut_(iConfig.getParameter<double>("relIso_cut")),
+  miniRelIso_cut_(iConfig.getParameter<double>("miniRelIso_cut")),
+  caloJet_DR_(iConfig.getParameter<double>("caloJet_DR"))
 {
+    // TrackAssociator parameters
+    edm::ParameterSet parameters = iConfig.getParameter<edm::ParameterSet>("TrackAssociatorParameters");
+    edm::ConsumesCollector iC = consumesCollector();
+    trackAssocParameters_.loadParameters( parameters, iC );
+
+    trackAssociator_.useDefaultPropagator();
+
     produces< pat::IsolatedTrackCollection > ();
 }
 
 pat::PATIsolatedTrackProducer::~PATIsolatedTrackProducer() {}
 
-
-void pat::PATIsolatedTrackProducer::produce(edm::StreamID stream, edm::Event& iEvent, const edm::EventSetup& iSetup) const {
+void pat::PATIsolatedTrackProducer::produce(edm::Event& iEvent, const edm::EventSetup& iSetup) {
 
     edm::Handle<pat::PackedCandidateCollection> pc_h;
     iEvent.getByToken( pc_, pc_h );
     const pat::PackedCandidateCollection *pc = pc_h.product();
 
-    auto outPtrP = std::make_unique<std::vector<pat::IsolatedTrack>>();
+    edm::Handle<pat::PackedCandidateCollection> lt_h;
+    iEvent.getByToken( lt_, lt_h );
+
+    edm::Handle<reco::TrackCollection> gt_h;
+    iEvent.getByToken( gt_, gt_h );
+    const reco::TrackCollection *generalTracks = gt_h.product();
+
+
+    edm::Handle<edm::Association<pat::PackedCandidateCollection> > gt2pc;
+    iEvent.getByToken(gt2pc_, gt2pc);
+
+    edm::Handle<edm::Association<pat::PackedCandidateCollection> > gt2lt;
+    iEvent.getByToken(gt2lt_, gt2lt);
+
+    edm::Handle<edm::Association<reco::PFCandidateCollection> > pc2pf;
+    iEvent.getByToken(pc2pf_, pc2pf);
+
+    edm::Handle<reco::CaloJetCollection> caloJets;
+    iEvent.getByToken(caloJets_, caloJets);
+
+    edm::Handle<edm::ValueMap<reco::DeDxData> > gt2dedx;
+    iEvent.getByToken(gt2dedx_, gt2dedx);
+
+    edm::Handle<reco::DeDxHitInfoAss> gt2dedxHitInfo;
+    iEvent.getByToken(gt2dedxHitInfo_, gt2dedxHitInfo);
+
+    edm::ESHandle<HcalChannelQuality> hcalQ_h;
+    iSetup.get<HcalChannelQualityRcd>().get("withTopo", hcalQ_h);
+    const HcalChannelQuality *hcalQ = hcalQ_h.product();
     
-    for(pat::PackedCandidateCollection::const_iterator pf_it = pc->begin(); pf_it != pc->end(); pf_it++){
+    edm::ESHandle<EcalChannelStatus> ecalS_h;
+    iSetup.get<EcalChannelStatusRcd>().get(ecalS_h);
+    const EcalChannelStatus *ecalS = ecalS_h.product();
 
-        if(pf_it->p4().pt() < pT_cut)
+    auto outPtrP = std::make_unique<std::vector<pat::IsolatedTrack>>();
+
+    //add general tracks
+    for(unsigned int igt=0; igt<generalTracks->size(); igt++){
+        const reco::Track &gentk = (*gt_h)[igt];
+        reco::TrackRef tkref = reco::TrackRef(gt_h, igt);
+        pat::PackedCandidateRef pcref = (*gt2pc)[tkref];
+        pat::PackedCandidateRef ltref = (*gt2lt)[tkref];
+        bool isInPackedCands = (pcref.isNonnull() && pcref.id()==pc_h.id());
+        bool isInLostTracks  = (ltref.isNonnull() && ltref.id()==lt_h.id());
+
+        // apparently this happens when a neutral pfcand is associated with a charged track
+        // if(isInPackedCands && isInLostTracks){
+        //     std::cout << "THIS SHOULDN'T HAPPEN " << iEvent.eventAuxiliary().event() << " " <<
+        //         tkref.key() << " " << tkref.get()->pt() << " " <<
+        //         pcref.key() << " " << pcref.get()->pt() << " " <<
+        //         ltref.key() << " " << ltref.get()->pt() << std::endl;
+        // }
+
+        LorentzVector p4;
+        pat::PackedCandidateRef refToCand;
+        int pdgId, charge, fromPV;
+        float dz, dxy, dzError, dxyError;
+        int pfCandInd; //to avoid counting packedPFCands in their own isolation
+
+        // get the four-momentum and charge
+        if(isInPackedCands && pcref.get()->charge() != 0){
+            p4 = pcref.get()->p4();
+            charge = pcref.get()->charge();
+            pfCandInd = pcref.key();
+        }else if(isInLostTracks){
+            p4 = ltref.get()->p4();
+            charge = ltref.get()->charge();
+            pfCandInd = -1;
+        }else{
+            double m = 0.13957018; //assume pion mass
+            double E = sqrt(m*m + gentk.p()*gentk.p());
+            p4.SetPxPyPzE(gentk.px(), gentk.py(), gentk.pz(), E);
+            charge = gentk.charge();
+            pfCandInd = -1;
+        }
+
+        if(p4.pt() < pT_cut_)
+            continue;
+        if(charge == 0)
             continue;
 
-        if(pf_it->fromPV()<=1 && fabs(pf_it->dz()) > dZ_cut)
+        // get the isolation of the track
+        pat::PFIsolation isolationDR03;
+        pat::PFIsolation miniIso;
+        getIsolation(p4, pc, pfCandInd, isolationDR03, miniIso);
+        
+        // isolation cut
+        if( p4.pt() < pT_cut_noIso_ && 
+            !(isolationDR03.chargedHadronIso() < absIso_cut_ ||
+              isolationDR03.chargedHadronIso()/p4.pt() < relIso_cut_ ||
+              miniIso.chargedHadronIso()/p4.pt() < miniRelIso_cut_))
             continue;
 
-        if(pf_it->charge() == 0)
+        // get the rest after the pt/iso cuts. Saves some runtime
+        if(isInPackedCands && pcref.get()->charge() != 0){
+            pdgId = pcref.get()->pdgId();
+            dz = pcref.get()->dz();
+            dxy = pcref.get()->dxy();
+            dzError = pcref.get()->dzError();
+            dxyError = pcref.get()->dxyError();
+            fromPV = pcref.get()->fromPV();
+            refToCand = pcref;
+        }else if(isInLostTracks){
+            pdgId = 0;
+            dz = ltref.get()->dz();
+            dxy = ltref.get()->dxy();
+            dzError = ltref.get()->dzError();
+            dxyError = ltref.get()->dxyError();
+            fromPV = -1;
+            refToCand = ltref;
+        }else{
+            pdgId = 0;
+            dz = gentk.dz();
+            dxy = gentk.dxy();
+            dzError = gentk.dzError();
+            dxyError = gentk.dxyError();
+            fromPV = -1;
+            refToCand = pat::PackedCandidateRef();  //NULL reference
+        }
+
+        float caloJetEm, caloJetHad;
+        getCaloJetEnergy(p4, caloJets.product(), caloJetEm, caloJetHad);
+
+        const reco::DeDxHitInfo* hitInfo = (*gt2dedxHitInfo)[tkref].get();
+        float dEdxPixel = getDeDx(hitInfo, true, false);
+        float dEdxStrip = (*gt2dedx)[tkref].dEdx(); // estimated strip dEdx is already stored in AOD
+
+        int trackQuality = gentk.qualityMask();
+
+        // get the associated ecal/hcal detectors
+        TrackDetMatchInfo trackDetInfo = getTrackDetMatchInfo(iEvent, iSetup, gentk);
+
+        // convert to specific hcal DetIds and fill status vector
+        std::vector<HcalDetId> crossedHcalIds;
+        std::vector<uint32_t> crossedHcalStatus;
+        for(auto const & did : trackDetInfo.crossedHcalIds){
+            crossedHcalIds.push_back(HcalDetId(did));
+            crossedHcalStatus.push_back(hcalQ->getValues(did.rawId())->getValue());
+        }
+        // fill ecal status vector
+        std::vector<EcalChannelStatusCode> crossedEcalStatus;
+        for(auto const & did : trackDetInfo.crossedEcalIds){
+            crossedEcalStatus.push_back(*(ecalS->find(did.rawId())));
+        }
+
+        outPtrP->push_back(pat::IsolatedTrack(isolationDR03, miniIso, caloJetEm, caloJetHad, p4,
+                                             charge, pdgId, dz, dxy, dzError, dxyError,
+                                             gentk.hitPattern(), dEdxStrip, dEdxPixel, fromPV, trackQuality, 
+                                             trackDetInfo.crossedEcalIds, crossedHcalIds,
+                                             crossedEcalStatus, crossedHcalStatus, refToCand));
+
+    }
+
+    // there are some number of pfcandidates with no associated track
+    // (mostly electrons, with a handful of muons)
+    // here we find these and store. Track-specific variables get some default values
+    for(unsigned int ipc=0; ipc<pc->size(); ipc++){
+        const pat::PackedCandidate& pfcand = pc->at(ipc);
+        pat::PackedCandidateRef pcref = pat::PackedCandidateRef(pc_h, ipc);
+        reco::PFCandidateRef pfref = (*pc2pf)[pcref];
+
+        // already counted if it has a track reference in the generalTracks collection
+        if(pfref.get()->trackRef().isNonnull() && pfref.get()->trackRef().id() == gt_h.id())
             continue;
 
-        pat::PackedCandidateRef pcref = pat::PackedCandidateRef(pc_h, (unsigned int)(pf_it - pc->begin()));
+        LorentzVector p4;
+        pat::PackedCandidateRef refToCand;
+        int pdgId, charge, fromPV;
+        float dz, dxy, dzError, dxyError;
 
-        // compute both standard and mini PFIsolation
-        // mini-isolation reference: https://hypernews.cern.ch/HyperNews/CMS/get/susy/1991.html
+        p4 = pfcand.p4();
+        charge = pfcand.charge();
+
+        if(p4.pt() < pT_cut_)
+            continue;
+        if(charge == 0)
+            continue;
+
+        // get the isolation of the track
+        pat::PFIsolation isolationDR03;
+        pat::PFIsolation miniIso;
+        getIsolation(p4, pc, ipc, isolationDR03, miniIso);
+        
+        // isolation cut
+        if( p4.pt() < pT_cut_noIso_ && 
+            !(isolationDR03.chargedHadronIso() < absIso_cut_ ||
+              isolationDR03.chargedHadronIso()/p4.pt() < relIso_cut_ ||
+              miniIso.chargedHadronIso()/p4.pt() < miniRelIso_cut_))
+            continue;
+
+        pdgId = pfcand.pdgId();
+        dz = pfcand.dz();
+        dxy = pfcand.dxy();
+        dzError = pfcand.dzError();
+        dxyError = pfcand.dxyError();
+        fromPV = pfcand.fromPV();
+        refToCand = pcref;
+
+        float caloJetEm, caloJetHad;
+        getCaloJetEnergy(p4, caloJets.product(), caloJetEm, caloJetHad);
+
+        // fill with default values
+        reco::HitPattern hp;
+        float dEdxPixel=-1, dEdxStrip=-1;
+        int trackQuality=0;
+        std::vector<DetId> ecalIds;
+        std::vector<HcalDetId> hcalIds;
+        std::vector<EcalChannelStatusCode> ecalStatus;
+        std::vector<uint32_t> hcalStatus;
+
+        outPtrP->push_back(pat::IsolatedTrack(isolationDR03, miniIso, caloJetEm, caloJetHad, p4,
+                                             charge, pdgId, dz, dxy, dzError, dxyError,
+                                             hp, dEdxStrip, dEdxPixel, fromPV, trackQuality, 
+                                             ecalIds, hcalIds, ecalStatus, hcalStatus, refToCand));
+    }
+
+    iEvent.put(std::move(outPtrP));
+}
+
+
+void pat::PATIsolatedTrackProducer::getIsolation(const LorentzVector& p4, 
+                                                const pat::PackedCandidateCollection *pc, int pc_idx,
+                                                pat::PFIsolation &iso, pat::PFIsolation &miniiso)
+{
         float chiso=0, nhiso=0, phiso=0, puiso=0;   // standard isolation
         float chmiso=0, nhmiso=0, phmiso=0, pumiso=0;  // mini isolation
-        float miniDR = std::max(miniIsoParams[0], std::min(miniIsoParams[1], miniIsoParams[2]/pf_it->p4().pt()));
-        for(pat::PackedCandidateCollection::const_iterator pf_it2 = pc->begin(); pf_it2 != pc->end(); pf_it2++){
-            if(pf_it == pf_it2)
+        float miniDR = std::max(miniIsoParams_[0], std::min(miniIsoParams_[1], miniIsoParams_[2]/p4.pt()));
+        for(pat::PackedCandidateCollection::const_iterator pf_it = pc->begin(); pf_it != pc->end(); pf_it++){
+            if(pf_it - pc->begin() == pc_idx)  //don't count itself
                 continue;
-            int id = std::abs(pf_it2->pdgId());
-            bool fromPV = (pf_it2->fromPV()>1 || fabs(pf_it2->dz()) < dZ_cut);
-            float pt = pf_it2->p4().pt();
-            float dr = deltaR(pf_it->p4(), pf_it2->p4());
+            int id = std::abs(pf_it->pdgId());
+            bool fromPV = (pf_it->fromPV()>1 || fabs(pf_it->dz()) < pfIsolation_DZ_);
+            float pt = pf_it->p4().pt();
+            float dr = deltaR(p4, pf_it->p4());
 
-            if(dr < dR_cut){
+            if(dr < pfIsolation_DR_){
                 // charged cands from PV get added to trackIso
                 if(id==211 && fromPV)
                     chiso += pt;
@@ -112,22 +383,95 @@ void pat::PATIsolatedTrackProducer::produce(edm::StreamID stream, edm::Event& iE
                     phmiso += pt;
             }
         }
-        
-        pat::PFIsolation isolationDR03(chiso, nhiso, phiso, puiso);
-        pat::PFIsolation miniIso(chmiso, nhmiso, phmiso, pumiso);
 
-        if(isolationDR03.chargedHadronIso() < absIso_cut ||
-           isolationDR03.chargedHadronIso()/pf_it->p4().pt() < relIso_cut ||
-           miniIso.chargedHadronIso()/pf_it->p4().pt() < miniRelIso_cut){
-            outPtrP->push_back(pat::IsolatedTrack(isolationDR03, miniIso, pf_it->p4(), 
-                                                  pf_it->charge(), pf_it->pdgId(), pf_it->dz(),
-                                                  pf_it->dxy(), pcref));
-        }
+        iso = pat::PFIsolation(chiso,nhiso,phiso,puiso);
+        miniiso = pat::PFIsolation(chmiso,nhmiso,phmiso,pumiso);
 
+}
+
+// get the estimated DeDx in either the pixels or strips (or both)
+float pat::PATIsolatedTrackProducer::getDeDx(const reco::DeDxHitInfo *hitInfo, bool doPixel, bool doStrip)
+{
+    if(hitInfo == NULL){
+        return -1;
     }
 
+    std::vector<float> charge_vec;
+    for(unsigned int ih=0; ih<hitInfo->size(); ih++){
 
-    iEvent.put(std::move(outPtrP));
+        // only use pixels. Am i doing this right?
+        // if(hitInfo->pixelCluster(ih) == nullptr) continue;
+        bool isPixel = (hitInfo->pixelCluster(ih) != nullptr);
+        bool isStrip = (hitInfo->stripCluster(ih) != nullptr);
+
+        if(isPixel && !doPixel) continue;
+        if(isStrip && !doStrip) continue;
+
+        // probably shouldn't happen
+        if(!isPixel && !isStrip) continue;
+
+        // shape selection for strips
+        if(isStrip && !DeDxTools::shapeSelection(*(hitInfo->stripCluster(ih))))
+            continue;
+        
+        float Norm=0;       
+        if(isPixel)
+            Norm = 3.61e-06; //compute the normalization factor to get the energy in MeV/mm
+        if(isStrip)
+            Norm = 3.61e-06 * 265;
+            
+        //save the dE/dx in MeV/mm to a vector.  
+        charge_vec.push_back(Norm*hitInfo->charge(ih)/hitInfo->pathlength(ih));           
+    }
+
+    int size = charge_vec.size();
+    float result = 0.0;
+
+    //build the harmonic 2 dE/dx estimator
+    float expo = -2;
+    for(int i=0; i<size; i++){
+        result += pow(charge_vec[i], expo);
+    }
+    result = (size>0) ? pow(result/size, 1./expo) : 0.0;
+
+    return result;
+}
+
+TrackDetMatchInfo pat::PATIsolatedTrackProducer::getTrackDetMatchInfo(const edm::Event& iEvent, 
+                                                                     const edm::EventSetup& iSetup,
+                                                                     const reco::Track& track)
+{
+    edm::ESHandle<MagneticField> bField;
+    iSetup.get<IdealMagneticFieldRecord>().get(bField);
+    FreeTrajectoryState initialState = trajectoryStateTransform::initialFreeState(track,&*bField);
+
+    // can't use the associate() using reco::Track directly, since 
+    // track->extra() is non-null but segfaults when trying to use it
+    return trackAssociator_.associate(iEvent, iSetup, trackAssocParameters_, &initialState);
+}
+
+void pat::PATIsolatedTrackProducer::getCaloJetEnergy(const LorentzVector& p4, const reco::CaloJetCollection* cJets,
+                                                     float &caloJetEm, float& caloJetHad)
+{
+    float nearestDR = 1.0;
+    int ind = -1;
+    for(unsigned int i=0; i<cJets->size(); i++){
+        float dR = deltaR(cJets->at(i).p4(), p4);
+        if(dR < caloJet_DR_ && dR < nearestDR){
+            nearestDR = dR;
+            ind = i;
+        }
+    }
+
+    if(ind==-1){
+        caloJetEm = 0;
+        caloJetHad = 0;
+    }else{
+        caloJetEm = cJets->at(ind).emEnergyInEB() + cJets->at(ind).emEnergyInEE() + cJets->at(ind).emEnergyInHF();
+        caloJetHad = cJets->at(ind).hadEnergyInHB() + cJets->at(ind).hadEnergyInHE() + 
+            cJets->at(ind).hadEnergyInHF() + cJets->at(ind).hadEnergyInHO();
+    }
+
 }
