+++ PhysicsTools/PatAlgos/python/slimming/slimming_cff.py
@@ -2,7 +2,6 @@ import FWCore.ParameterSet.Config as cms
 
 from PhysicsTools.PatAlgos.slimming.packedPFCandidates_cff import *
 from PhysicsTools.PatAlgos.slimming.isolatedTracks_cfi import *
-from PhysicsTools.PatAlgos.slimming.stoppedTracks_cfi import *
 from PhysicsTools.PatAlgos.slimming.lostTracks_cfi import *
 from PhysicsTools.PatAlgos.slimming.offlineSlimmedPrimaryVertices_cfi import *
 from PhysicsTools.PatAlgos.slimming.primaryVertexAssociation_cfi import *
@@ -25,9 +24,8 @@ from HeavyFlavorAnalysis.Onia2MuMu.OniaPhotonConversionProducer_cfi import Photo
 
 slimmingTask = cms.Task(
     packedPFCandidatesTask,
-    isolatedTracks,
     lostTracks,
-    stoppedTracks,
+    isolatedTracks,
     offlineSlimmedPrimaryVertices,
     primaryVertexAssociation,
     genParticlesTask,
